+++ examples/3d/complex_one_color.rs
@@ -1,18 +1,22 @@
 use bevy::prelude::*;
 use bevy_svg::prelude::*;
 
+#[path = "../common/lib.rs"]
+mod common;
+
 fn main() {
     App::new()
         .insert_resource(Msaa { samples: 4 })
         .add_plugins(DefaultPlugins.set(WindowPlugin {
             window: WindowDescriptor {
-                title: "complex_one_color".to_string(),
-                width: 400.0,
-                height: 400.0,
+                title: "3d_complex_one_color".to_string(),
+                width: 600.0,
+                height: 600.0,
                 ..Default::default()
             },
             ..Default::default()
         }))
+        .add_plugin(common::CommonPlugin)
         .add_plugin(bevy_svg::prelude::SvgPlugin)
         .add_startup_system(setup)
         .run();
@@ -25,8 +29,8 @@ fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
         svg,
         origin: Origin::Center,
         transform: Transform {
-            translation: Vec3::new(0.0, 0.0, -1.0),
-            scale: Vec3::new(0.005, 0.005, 1.0),
+            translation: Vec3::new(0.0, 0.0, -600.0),
+            scale: Vec3::new(2.0, 2.0, 1.0),
             ..Default::default()
         },
         ..Default::default()