+++ examples/3d/twinkle.rs
@@ -1,18 +1,22 @@
 use bevy::prelude::*;
 use bevy_svg::prelude::*;
 
+#[path = "../common/lib.rs"]
+mod common;
+
 fn main() {
     App::new()
         .insert_resource(Msaa { samples: 4 })
         .add_plugins(DefaultPlugins.set(WindowPlugin {
             window: WindowDescriptor {
-                title: "twinkle".to_string(),
-                width: 400.0,
-                height: 400.0,
+                title: "3d_twinkle".to_string(),
+                width: 600.0,
+                height: 600.0,
                 ..Default::default()
             },
             ..Default::default()
         }))
+        .add_plugin(common::CommonPlugin)
         .add_plugin(bevy_svg::prelude::SvgPlugin)
         .add_startup_system(setup)
         .run();